requirements.txt
numpy==1.16.2
matplotlib==3.1.1
tqdm==4.36.1
scipy==1.3.1
easydict==1.9
opencv-python==4.1.1.26
Pillow==6.1.0
mxnet==1.5.1.post0
tensorboardX==2.0
torch==1.1.0
torchvision==0.3.0
scikit-learn==0.22.1
xlwt==1.3.0

bcolz==1.2.1
mxnet_cu90
